refactor database class from launcher class
