--- packages/fireproof/src/index.js
@@ -1,6 +1,0 @@
-import { Fireproof } from './fireproof.js'
-import { DbIndex as Index } from './db-index.js'
-import { Listener } from './listener.js'
-import { Hydrator } from './hydrator.js'
-
-export { Fireproof, Index, Listener, Hydrator }
